+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Features
+
+- Add descriptive tooltips for Input, Reference, Auxiliary and Media file selectors in the main webview
+
 ## [0.33.0] - 2025-08-19 🎂 Birthday Edition
 
 ### Features